GroupKV: Move test code into its own file in tests
