+++ tests/CMakeLists.txt
@@ -117,6 +117,7 @@ llama_target_and_test(test-quantize-perf.cpp)
 llama_target_and_test(test-sampling.cpp)
 llama_target_and_test(test-chat-template.cpp)
 llama_target_and_test(test-chat-template-chaton.cpp)
+llama_target_and_test(test-chaton-groupkv.cpp)
 
 llama_target_and_test(test-grammar-parser.cpp)
 llama_target_and_test(test-llama-grammar.cpp)
